Controller files reorganization
