+++ DomeController/src/controller.h
@@ -13,9 +13,9 @@
 #include "configuration.h"
 
 #if MOTOR_CONTROLLER == MOTOR_CONTROLLER_BTS7960
-    #include "BTS7960Controller.h"
+    #include "controllers/BTS7960Controller.h"
 #elif MOTOR_CONTROLLER == MOTOR_CONTROLLER_SHIELDMD10
-    #include "SHIELDMD10Controller.h"
+    #include "controllers/SHIELDMD10Controller.h"
 #endif
 
 // Define a pointer to a function for checking controller/flap
