+++ DomeController/src/controllers/BTS7960Controller.h
