+++ DomeController/src/main.cpp
@@ -32,9 +32,9 @@ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 #include "configuration.h"
 
 #if MOTOR_CONTROLLER == MOTOR_CONTROLLER_BTS7960
-    #include "BTS7960Controller.h"
+    #include "controllers/BTS7960Controller.h"
 #elif MOTOR_CONTROLLER == MOTOR_CONTROLLER_SHIELDMD10
-    #include "SHIELDMD10Controller.h"
+    #include "controllers/SHIELDMD10Controller.h"
 #endif
 
 #include "serialCommand.h"
